+++ packages/manager/src/components/Uploaders/FileUpload.styles.ts
