+++ packages/manager/src/components/Uploaders/FileUploader/FileUploader.stories.tsx
@@ -1,0 +1,4 @@
+import * as React from 'react';
+
+import { FileUploader } from './FileUploader';
+import type { Meta, StoryObj } from '@storybook/react';
