+++ packages/manager/src/components/Uploaders/ObjectUploader/ObjectUploader.tsx
@@ -1,4 +1,5 @@
 import { getObjectURL } from '@linode/api-v4/lib/object-storage';
+import { AxiosProgressEvent } from 'axios';
 import { useSnackbar } from 'notistack';
 import * as React from 'react';
 import { FileRejection, useDropzone } from 'react-dropzone';
@@ -10,8 +11,8 @@ import { updateBucket } from 'src/queries/objectStorage';
 import { sendObjectsQueuedForUploadEvent } from 'src/utilities/analytics';
 import { readableBytes } from 'src/utilities/unitConversions';
 
-import { uploadObject } from '../requests';
-import { FileUpload } from './FileUpload';
+import { uploadObject } from '../../../features/ObjectStorage/requests';
+import { FileUpload } from '../FileUpload';
 import {
   StyledDropZoneContent,
   StyledDropZoneCopy,
@@ -26,13 +27,24 @@ import {
   curriedObjectUploaderReducer,
   defaultState,
   pathOrFileName,
-} from './reducer';
-import { AxiosProgressEvent } from 'axios';
+} from '../reducer';
 
 interface Props {
+  /**
+   * The Object Storage bucket to upload to.
+   */
   bucketName: string;
+  /**
+   * The Object Storage cluster to upload to.
+   */
   clusterId: string;
+  /**
+   * A function that is called when an Object is uploaded successfully so we can manually update our local store to reflect the upload
+   */
   maybeAddObjectToTable: (path: string, sizeInBytes: number) => void;
+  /**
+   * The Object Storage prefix (path) to upload to.
+   */
   prefix: string;
 }
