+++ packages/manager/src/components/Uploaders/reducer.ts
